EZP-24664: adapt Elasticsearch engine
