+++ eZ/Publish/Core/Search/Elasticsearch/Handler.php
@@ -13,7 +13,6 @@ namespace eZ\Publish\Core\Search\Elasticsearch;
 
 use eZ\Publish\SPI\Search\Handler as HandlerInterface;
 use eZ\Publish\SPI\Search\Content\Handler as ContentSearchHandler;
-use eZ\Publish\SPI\Search\Content\Location\Handler as LocationSearchHandler;
 
 /**
  * The main handler for the Elasticsearch Search Engine.
@@ -25,26 +24,13 @@ class Handler implements HandlerInterface
      */
     protected $contentSearchHandler;
 
-    /**
-     * @var \eZ\Publish\SPI\Search\Content\Location\Handler
-     */
-    protected $locationSearchHandler;
-
-    public function __construct(
-        ContentSearchHandler $contentSearchHandler,
-        LocationSearchHandler $locationSearchHandler
-    ) {
+    public function __construct(ContentSearchHandler $contentSearchHandler)
+    {
         $this->contentSearchHandler = $contentSearchHandler;
-        $this->locationSearchHandler = $locationSearchHandler;
     }
 
     public function contentSearchHandler()
     {
         return $this->contentSearchHandler;
-    }
-
-    public function locationSearchHandler()
-    {
-        return $this->locationSearchHandler;
     }
 }
